sms_data_gen/requirements.txt
pillow==12.1.0
wheel==0.45.1
